requirements.txt
astroid==3.3.8
dill==0.3.9
Faker==36.1.1
iniconfig==2.0.0
isort==6.0.0
mccabe==0.7.0
packaging==24.2
platformdirs==4.3.6
pluggy==1.5.0
pylint==3.3.4
pytest==8.3.4
pytest-pylint==0.21.0
tomlkit==0.13.2
tzdata==2025.1
python-dotenv==0.21.0
